+++ src/onyx/log/commands/volunteer_for_task.clj
@@ -18,13 +18,6 @@
         completed (get-in replica [:completions job])]
     (filter identity (second (diff completed tasks)))))
 
-(defmethod select-task :onyx.task-scheduler/greedy
-  [replica job peer-id]
-  (->> (get-in replica [:tasks job])
-       (incomplete-tasks replica job)
-       (common/active-tasks-only replica)
-       (first)))
-
 (defmethod select-task :onyx.task-scheduler/round-robin
   [replica job peer-id]
   (let [allocations (get-in replica [:allocations job])]