+++ test/onyx/log/leave_cluster_test.clj
@@ -51,8 +51,8 @@
                   :pairs {:a :b :b :c :c :d :d :a} 
                   :peers [:a :b :c :d]
                   :jobs [:j1 :j2]
-                  :task-schedulers {:j1 :onyx.task-scheduler/greedy
-                                    :j2 :onyx.task-scheduler/greedy} 
+                  :task-schedulers {:j1 :onyx.task-scheduler/round-robin
+                                    :j2 :onyx.task-scheduler/round-robin} 
                   :tasks {:j1 [:t1] :j2 [:t2]}
                   :allocations {:j1 {:t1 [:a :b]}
                                 :j2 {:t2 [:c :d]}}})
@@ -76,8 +76,8 @@
                   :pairs {:a :b :b :c :c :a} 
                   :peers [:a :b :c]
                   :jobs [:j1 :j2]
-                  :task-schedulers {:j1 :onyx.task-scheduler/greedy
-                                    :j2 :onyx.task-scheduler/greedy} 
+                  :task-schedulers {:j1 :onyx.task-scheduler/round-robin
+                                    :j2 :onyx.task-scheduler/round-robin} 
                   :tasks {:j1 [:t1] :j2 [:t2]}
                   :allocations {:j1 {:t1 [:a :b]} :j2 {:t2 [:c]}}})
 
@@ -92,8 +92,8 @@
                   :pairs {:a :b :b :c :c :a} 
                   :peers [:a :b :c]
                   :jobs [:j1 :j2]
-                  :task-schedulers {:j1 :onyx.task-scheduler/greedy
-                                    :j2 :onyx.task-scheduler/greedy} 
+                  :task-schedulers {:j1 :onyx.task-scheduler/round-robin
+                                    :j2 :onyx.task-scheduler/round-robin} 
                   :tasks {:j1 [:t1] :j2 [:t2]}
                   :allocations {:j1 {:t1 [:a :b]} :j2 {:t2 [:c]}}
                   :sealing-tasks {:t2 :c}})